broke index.js down into seperate files
