+++ data/dbConfig.js
@@ -1,0 +1,4 @@
+const knex = require('knex');
+const dbConfig = require('../knexfile');
+
+module.exports = knex(dbConfig.development)
