+++ package.json
@@ -9,6 +9,7 @@
 	"author": "Lambda School",
 	"license": "ISC",
 	"dependencies": {
+		"cors": "^2.8.5",
 		"express": "^4.16.4",
 		"helmet": "^3.15.0",
 		"knex": "^0.16.3",
